moved tests to a separate package
