+++ .travis.yml
@@ -1,3 +1,6 @@
 language: java
-
-script: ant
+notifications:
+  # Email notifications are disabled to not annoy anybody.
+  # See http://about.travis-ci.org/docs/user/build-configuration/ to learn more
+  # about configuring notification recipients and more.
+  email: false
