--- build.properties
@@ -1,4 +1,0 @@
-path.variable.maven_repository=C\:\\Users\\Marko\\.m2\\repository
-jdk.home.1.7=C\:/Program Files/Java/jdk1.7.0_80
-idea.home=C\:\\Program Files (x86)\\JetBrains\\IntelliJ IDEA 14.1.5
-javac2.instrumentation.includeJavaRuntime=false
