--- module_learningjava.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project name="module_learningjava" default="compile.module.learningjava">
-  <dirname property="module.learningjava.basedir" file="${ant.file.module_learningjava}"/>
-  
-  <property name="module.jdk.home.learningjava" value="${project.jdk.home}"/>
-  <property name="module.jdk.bin.learningjava" value="${project.jdk.bin}"/>
-  <property name="module.jdk.classpath.learningjava" value="${project.jdk.classpath}"/>
-  
-  <property name="compiler.args.learningjava" value="-encoding UTF-8 -source 1.7 ${compiler.args}"/>
-  
-  <property name="learningjava.output.dir" value="${module.learningjava.basedir}/out/production/LearningJava"/>
-  <property name="learningjava.testoutput.dir" value="${module.learningjava.basedir}/out/test/LearningJava"/>
-  
-  <path id="learningjava.module.bootclasspath">
-    <!-- Paths to be included in compilation bootclasspath -->
-  </path>
-  
-  <path id="learningjava.module.production.classpath">
-    <path refid="${module.jdk.classpath.learningjava}"/>
-    <pathelement location="${idea.home}/lib/hamcrest-core-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/hamcrest-library-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/junit-4.11.jar"/>
-  </path>
-  
-  <path id="learningjava.runtime.production.module.classpath">
-    <pathelement location="${learningjava.output.dir}"/>
-    <pathelement location="${idea.home}/lib/hamcrest-core-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/hamcrest-library-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/junit-4.11.jar"/>
-  </path>
-  
-  <path id="learningjava.module.classpath">
-    <path refid="${module.jdk.classpath.learningjava}"/>
-    <pathelement location="${learningjava.output.dir}"/>
-    <pathelement location="${idea.home}/lib/hamcrest-core-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/hamcrest-library-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/junit-4.11.jar"/>
-  </path>
-  
-  <path id="learningjava.runtime.module.classpath">
-    <pathelement location="${learningjava.testoutput.dir}"/>
-    <pathelement location="${learningjava.output.dir}"/>
-    <pathelement location="${idea.home}/lib/hamcrest-core-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/hamcrest-library-1.3.jar"/>
-    <pathelement location="${idea.home}/lib/junit-4.11.jar"/>
-  </path>
-  
-  
-  <patternset id="excluded.from.module.learningjava">
-    <patternset refid="ignored.files"/>
-  </patternset>
-  
-  <patternset id="excluded.from.compilation.learningjava">
-    <patternset refid="excluded.from.module.learningjava"/>
-  </patternset>
-  
-  <path id="learningjava.module.sourcepath">
-    <dirset dir="${module.learningjava.basedir}">
-      <include name="src"/>
-    </dirset>
-  </path>
-  
-  
-  <target name="compile.module.learningjava" depends="compile.module.learningjava.production,compile.module.learningjava.tests" description="Compile module LearningJava"/>
-  
-  <target name="compile.module.learningjava.production" depends="register.custom.compilers" description="Compile module LearningJava; production classes">
-    <mkdir dir="${learningjava.output.dir}"/>
-    <javac2 destdir="${learningjava.output.dir}" debug="${compiler.debug}" nowarn="${compiler.generate.no.warnings}" memorymaximumsize="${compiler.max.memory}" fork="true" executable="${module.jdk.bin.learningjava}/javac">
-      <compilerarg line="${compiler.args.learningjava}"/>
-      <bootclasspath refid="learningjava.module.bootclasspath"/>
-      <classpath refid="learningjava.module.production.classpath"/>
-      <src refid="learningjava.module.sourcepath"/>
-      <patternset refid="excluded.from.compilation.learningjava"/>
-    </javac2>
-    
-    <copy todir="${learningjava.output.dir}">
-      <fileset dir="${module.learningjava.basedir}/src">
-        <patternset refid="compiler.resources"/>
-        <type type="file"/>
-      </fileset>
-    </copy>
-  </target>
-  
-  <target name="compile.module.learningjava.tests" depends="register.custom.compilers,compile.module.learningjava.production" description="compile module LearningJava; test classes" unless="skip.tests"/>
-  
-  <target name="clean.module.learningjava" description="cleanup module">
-    <delete dir="${learningjava.output.dir}"/>
-    <delete dir="${learningjava.testoutput.dir}"/>
-  </target>
-</project>
