+++ test/kru/collections/testing/HashMapTest.java
@@ -1,4 +1,4 @@
-package kru.collections.tests;
+package kru.collections.testing;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
